Conflicts:
	cocos/scripting/js-bindings/proj.ios_mac/cocos2d_js_bindings.xcodeproj/project.pbxproj
	tests/js-tests/project.json
	tests/js-tests/src/GUITest/UIWebViewTest/UIWebViewTest.js
	tools/tojs/genbindings.py

+++ cocos/2d/CCFontFNT.cpp
@@ -298,7 +298,7 @@ std::set<unsigned int>* BMFontConfiguration::parseConfigFile(const std::string& 
     int parseCount = 0;
     while (next)
     {
-        lineLength = next - base;
+        lineLength = ((int)(next - base));
         memcpy(line, contents + parseCount, lineLength);
         line[lineLength] = 0;
